change int to long

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
 
     public static void main(String[] args) {
-        SalesManager salesManager = new SalesManager(new int[]{345,78,50});
+        SalesManager salesManager = new SalesManager(new long[]{345,78,50});
         System.out.println(salesManager.max());
     }
 }
